chore: split communication-related functionality from mempool and mempool-types modules
commit-id:8c4b0394
