+++ crates/mempool/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod communication;
 pub mod mempool;
 pub(crate) mod priority_queue;
