+++ crates/mempool/src/mempool.rs
@@ -1,17 +1,12 @@
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::HashMap;
 
-use async_trait::async_trait;
 use starknet_api::core::ContractAddress;
 use starknet_api::transaction::TransactionHash;
-use starknet_mempool_infra::component_definitions::ComponentRequestHandler;
-use starknet_mempool_infra::component_server::ComponentServer;
 use starknet_mempool_types::errors::MempoolError;
 use starknet_mempool_types::mempool_types::{
-    Account, AccountState, MempoolInput, MempoolRequest, MempoolRequestAndResponseSender,
-    MempoolResponse, MempoolResult, ThinTransaction,
+    Account, AccountState, MempoolInput, MempoolResult, ThinTransaction,
 };
-use tokio::sync::mpsc::Receiver;
 
 use crate::priority_queue::TransactionPriorityQueue;
 
@@ -99,48 +94,4 @@ impl Mempool {
     ) -> MempoolResult<()> {
         todo!()
     }
-}
-
-/// Wraps the mempool to enable inbound async communication from other components.
-pub struct MempoolCommunicationWrapper {
-    mempool: Mempool,
-}
-
-impl MempoolCommunicationWrapper {
-    pub fn new(mempool: Mempool) -> Self {
-        MempoolCommunicationWrapper { mempool }
-    }
-
-    fn add_tx(&mut self, mempool_input: MempoolInput) -> MempoolResult<()> {
-        self.mempool.add_tx(mempool_input.tx, mempool_input.account)
-    }
-
-    fn get_txs(&mut self, n_txs: usize) -> MempoolResult<Vec<ThinTransaction>> {
-        self.mempool.get_txs(n_txs)
-    }
-}
-
-#[async_trait]
-impl ComponentRequestHandler<MempoolRequest, MempoolResponse> for MempoolCommunicationWrapper {
-    async fn handle_request(&mut self, request: MempoolRequest) -> MempoolResponse {
-        match request {
-            MempoolRequest::AddTransaction(mempool_input) => {
-                MempoolResponse::AddTransaction(self.add_tx(mempool_input))
-            }
-            MempoolRequest::GetTransactions(n_txs) => {
-                MempoolResponse::GetTransactions(self.get_txs(n_txs))
-            }
-        }
-    }
-}
-
-type MempoolCommunicationServer =
-    ComponentServer<MempoolCommunicationWrapper, MempoolRequest, MempoolResponse>;
-
-pub fn create_mempool_server(
-    mempool: Mempool,
-    rx_mempool: Receiver<MempoolRequestAndResponseSender>,
-) -> MempoolCommunicationServer {
-    let mempool_communication_wrapper = MempoolCommunicationWrapper::new(mempool);
-    ComponentServer::new(mempool_communication_wrapper, rx_mempool)
 }
