+++ crates/mempool_types/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod communication;
 pub mod errors;
 pub mod mempool_types;
